finished issue functionality, missing some styling
When signed in as instructor and you go to localhost:3000 you can see the unchanged home page.

Some versions of the edit team isn't up to date with the new html changes. (like when you first create a team, or when you update a team).

Removing a user from a team should refresh right away as you click.